refactor: code splitting across playlists, video components
