+++ src/components/History/History.jsx
@@ -1,11 +1,13 @@
 import React from "react";
 import { useVideo } from "../../contexts";
+import { useDocumentTitle } from "../../utils";
 import { VideoListing, PlayListHeading } from "../index";
 
 function History() {
   const {
     state: { history },
   } = useVideo();
+  useDocumentTitle("History");
 
   return (
     <div>
