+++ src/components/Like/Like.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 import { useVideo } from "../../contexts";
+import { useDocumentTitle } from "../../utils";
 import { VideoListing, PlayListHeading } from "../index";
 
 function Like() {
   const {
     state: { likedVideos },
   } = useVideo();
-
+  useDocumentTitle("Liked");
   return (
     <div>
       <PlayListHeading
