+++ src/components/Playlists/PlayLists.jsx
@@ -1,13 +1,15 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { useVideo } from "../../contexts";
-import { getFilteredList } from "../../utils";
+import { getFilteredList, useDocumentTitle } from "../../utils";
+import PlayListView from "./PlayListView";
 import "./PlayLists.css";
 
 function PlayLists() {
   const {
     state: { playlists, likedVideos, savedVideos },
   } = useVideo();
+  useDocumentTitle("Playlists");
   const updatedPlaylists = getFilteredList(playlists);
   return (
     <>
@@ -42,35 +44,5 @@ function PlayLists() {
     </>
   );
 }
-
-const PlayListView = ({ name, videoList }) => (
-  <>
-    {videoList.length > 0 ? (
-      <div className="playlist--details">
-        <div className="playlist__thumbnail">
-          <img
-            className="img--responsive"
-            src={videoList[0].thumbnailUrl}
-            alt=""
-          />
-          <div className="playlist--overlay">
-            <div>{videoList.length}</div>
-            <img
-              className="img--xs"
-              src="https://img.icons8.com/fluent-systems-filled/48/ffffff/video-playlist.png"
-              alt=""
-            />
-          </div>
-        </div>
-        <div className="playlist__name">
-          <p>{name}</p>
-          <p className="text--gray body--md">{videoList.length} videos</p>
-        </div>
-      </div>
-    ) : (
-      <></>
-    )}
-  </>
-);
 
 export default PlayLists;
