+++ src/components/Playlists/PlaylistDetail/PlayListDetail.css
@@ -1,6 +1,10 @@
-.playlist__heading{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    align-items: center;
-}   
+.playlist__heading {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.pointer {
+  cursor: pointer;
+}
