+++ src/components/Video/VideoListing/VideoListing.css
@@ -72,6 +72,7 @@
   border-radius: 50%;
   width: 1.5rem;
   height: 1.5rem;
+  cursor: pointer;
 }
 
 .badge__icon--align:hover {
