+++ src/components/index.js
@@ -1,13 +1,13 @@
 export { default as Video } from "./Video/Video";
 export { default as VideoListing } from "./Video/VideoListing/VideoListing";
 export { default as VideoDetail } from "./Video/VideoDetail/VideoDetail";
-export {
-  default as PlayListDetail,
-  PlayListHeading,
-} from "./Playlists/PlaylistDetail/PlayListDetail";
+export { default as PlayListDetail } from "./Playlists/PlaylistDetail/PlayListDetail";
+export { default as PlayListHeading } from "./Playlists/PlaylistDetail/PlayListHeading";
 export { default as PlayLists } from "./Playlists/PlayLists";
 export { default as Like } from "./Like/Like";
 export { default as Saved } from "./Saved/Saved";
 export { default as History } from "./History/History";
 export { default as Toast } from "./Toast/Toast";
+export { default as NotFound } from "./NotFound/NotFound";
 export { handleToast } from "./Toast/Toast";
+export { closeToast } from "./Toast/Toast";
